fixed some more issues
