+++ reducer-todo/src/Components/Todo.js
@@ -4,7 +4,7 @@ import './Todo.css';
 export default function Todo(props) {
     return (
         <div>
-            <p> {props.item.item}</p>
+            <p className="todos" > {props.item.item}</p>
         </div>
     )
 }
