+++ reducer-todo/src/reducers/reducer.js
@@ -18,15 +18,6 @@ export const initialState = {
     }]
 };
         
-// const action1 = {
-//     type: 'ADD_TODO',
-//     payload: {
-//         item: 'Enjoy React from now on',
-//         id: 1234512345,
-//         completed: false
-//     }
-// }
-
 export const reducer = (state, action) => {
     switch (action.type) {
         
@@ -38,33 +29,31 @@ export const reducer = (state, action) => {
                 todos: [ ...state.todos, newTodo]
             };
             
-        case 'TOGGLE_COMPLETED':
-            const newState = {
-                ...state,
-                todos: [...state.todos]
-            }
+        // case 'TOGGLE_COMPLETED':
+        //     const newState = {
+        //         ...state,
+        //         todos: [...state.todos]
+        //     }
                  
-            return newState.todos.map(todo => {
-                if (todo.item === action.payload.item) {
-                    todo.completed = action.payload;
-                }
-                return newState;
-            });
+        //     return newState.todos.map(todo => {
+        //         if (todo.item === action.payload.item) {
+        //             todo.completed = action.payload;
+        //         }
+        //         return newState;
+        //     });
             
-        case 'DELETE_TODO':
-            const newState1 = {
-                ...state,
-                todos: [...state.todos]
-            }
-            return newState1.todos.map(todo => {
-                if (todo.completed === action.payload) {
-                    newState.todos.splice(newState.todos.indexOf(todo), 1);
-                }
-                return newState1;
-            });
+        // case 'DELETE_TODO':
+        //     const newState1 = {
+        //         ...state,
+        //         todos: [...state.todos]
+        //     }
+        //     return newState1.todos.map(todo => {
+        //         if (todo.completed === action.payload) {
+        //             newState.todos.splice(newState.todos.indexOf(todo), 1);
+        //         }
+        //         return newState1;
+        //     });
             
         default: return state;
     }
 };
-
-// export default reducer;
